requirements.txt
dash==2.9.3
dash-bootstrap-components==1.4.1
pandas==2.0.0
plotly==5.14.1
openai==1.1.1
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.10.1
